phan quyen : edit diem

+++ QuanLyDiem/routes/web.php
@@ -46,32 +46,32 @@ Route::group(['middleware' => ['App\Http\Middleware\checkPermission']], function
 
     //Mon hoc co ban
     Route::name('monManage.')->group(function () {
-        Route::get('/mon', 'App\Http\Controllers\MonController@index')->name('indexMon');
-        Route::get('/mon/create', 'App\Http\Controllers\MonController@create')->name('createMon');
-        Route::post('/mon/store', 'App\Http\Controllers\MonController@store')->name('storeMon');
-        Route::get('/mon/edit/{mamon}', 'App\Http\Controllers\MonController@edit')->name('editMon');
-        Route::post('/mon/update', 'App\Http\Controllers\MonController@update')->name('updateMon');
-        Route::delete('/mon/delete/{mamon}', 'App\Http\Controllers\MonController@delete')->name('deleteMon');
-    })->middleware('checkloginadmin::class');
+        Route::get('/mon', 'App\Http\Controllers\MonController@index')->name('indexMon')->middleware('checkloginadmin::class');
+        Route::get('/mon/create', 'App\Http\Controllers\MonController@create')->name('createMon')->middleware('checkloginadmin::class');
+        Route::post('/mon/store', 'App\Http\Controllers\MonController@store')->name('storeMon')->middleware('checkloginadmin::class');
+        Route::get('/mon/edit/{mamon}', 'App\Http\Controllers\MonController@edit')->name('editMon')->middleware('checkloginadmin::class');
+        Route::post('/mon/update', 'App\Http\Controllers\MonController@update')->name('updateMon')->middleware('checkloginadmin::class');
+        Route::delete('/mon/delete/{mamon}', 'App\Http\Controllers\MonController@delete')->name('deleteMon')->middleware('checkloginadmin::class');
+    });
 
     // Nien khoa
     Route::name('nienkhoaManage.')->group(function () {
-        Route::get('/nienkhoa', 'App\Http\Controllers\NienKhoaController@index')->name('indexNienKhoa');
-        Route::get('/nienkhoa/create', 'App\Http\Controllers\NienKhoaController@create')->name('createNienKhoa');
-        Route::post('/nienkhoa/store', 'App\Http\Controllers\NienKhoaController@store')->name('storeNienKhoa');
-        Route::get('/nienkhoa/edit/{manienkhoa}', 'App\Http\Controllers\NienKhoaController@edit')->name('editNienKhoa');
-        Route::post('/nienkhoa/update', 'App\Http\Controllers\NienKhoaController@update')->name('updateNienKhoa');
-        Route::delete('/nienkhoa/delete/{manienkhoa}', 'App\Http\Controllers\NienKhoaController@delete')->name('deleteNienKhoa');
+        Route::get('/nienkhoa', 'App\Http\Controllers\NienKhoaController@index')->name('indexNienKhoa')->middleware('checkloginadmin::class');
+        Route::get('/nienkhoa/create', 'App\Http\Controllers\NienKhoaController@create')->name('createNienKhoa')->middleware('checkloginadmin::class');
+        Route::post('/nienkhoa/store', 'App\Http\Controllers\NienKhoaController@store')->name('storeNienKhoa')->middleware('checkloginadmin::class');
+        Route::get('/nienkhoa/edit/{manienkhoa}', 'App\Http\Controllers\NienKhoaController@edit')->name('editNienKhoa')->middleware('checkloginadmin::class');
+        Route::post('/nienkhoa/update', 'App\Http\Controllers\NienKhoaController@update')->name('updateNienKhoa')->middleware('checkloginadmin::class');
+        Route::delete('/nienkhoa/delete/{manienkhoa}', 'App\Http\Controllers\NienKhoaController@delete')->name('deleteNienKhoa')->middleware('checkloginadmin::class');
     })->middleware('checkloginadmin::class');
 
     // Lop
     Route::name('lopManage.')->group(function () {
-        Route::get('/lop', 'App\Http\Controllers\LopController@index')->name('indexLop');
-        Route::get('/lop/create', 'App\Http\Controllers\LopController@create')->name('createLop');
-        Route::post('/lop/store', 'App\Http\Controllers\LopController@store')->name('storeLop');
-        Route::get('/lop/edit/{malop}', 'App\Http\Controllers\LopController@edit')->name('editLop');
-        Route::post('/lop/update', 'App\Http\Controllers\LopController@update')->name('updateLop');
-        Route::delete('/lop/delete/{malop}', 'App\Http\Controllers\LopController@delete')->name('deleteLop');
+        Route::get('/lop', 'App\Http\Controllers\LopController@index')->name('indexLop')->middleware('checkloginadmin::class');
+        Route::get('/lop/create', 'App\Http\Controllers\LopController@create')->name('createLop')->middleware('checkloginadmin::class');
+        Route::post('/lop/store', 'App\Http\Controllers\LopController@store')->name('storeLop')->middleware('checkloginadmin::class');
+        Route::get('/lop/edit/{malop}', 'App\Http\Controllers\LopController@edit')->name('editLop')->middleware('checkloginadmin::class');
+        Route::post('/lop/update', 'App\Http\Controllers\LopController@update')->name('updateLop')->middleware('checkloginadmin::class');
+        Route::delete('/lop/delete/{malop}', 'App\Http\Controllers\LopController@delete')->name('deleteLop')->middleware('checkloginadmin::class');
     })->middleware('checkloginadmin::class');
 
     // Quản lý thông tin học sinh
@@ -104,27 +104,27 @@ Route::group(['middleware' => ['App\Http\Middleware\checkPermission']], function
 
     //Mon Hoc
     Route::name('monhocManage.')->group(function () {
-        Route::get('/monhoc', 'App\Http\Controllers\MonHocController@index')->name('indexMonHoc');
-        Route::get('/monhoc/create', 'App\Http\Controllers\MonHocController@create')->name('createMonHoc');
-        Route::post('/monhoc/store', 'App\Http\Controllers\MonHocController@store')->name('storeMonHoc');
-        Route::get('/monhoc/edit/{mamonhoc}', 'App\Http\Controllers\MonHocController@edit')->name('editMonHoc');
-        Route::post('/monhoc/update', 'App\Http\Controllers\MonHocController@update')->name('updateMonHoc');
-        Route::delete('/monhoc/delete/{mamonhoc}', 'App\Http\Controllers\MonHocController@delete')->name('deleteMonHoc');
-    })->middleware('checkloginadmin::class');
+        Route::get('/monhoc', 'App\Http\Controllers\MonHocController@index')->name('indexMonHoc')->middleware('checkloginadmin::class');
+        Route::get('/monhoc/create', 'App\Http\Controllers\MonHocController@create')->name('createMonHoc')->middleware('checkloginadmin::class');
+        Route::post('/monhoc/store', 'App\Http\Controllers\MonHocController@store')->name('storeMonHoc')->middleware('checkloginadmin::class');
+        Route::get('/monhoc/edit/{mamonhoc}', 'App\Http\Controllers\MonHocController@edit')->name('editMonHoc')->middleware('checkloginadmin::class');
+        Route::post('/monhoc/update', 'App\Http\Controllers\MonHocController@update')->name('updateMonHoc')->middleware('checkloginadmin::class');
+        Route::delete('/monhoc/delete/{mamonhoc}', 'App\Http\Controllers\MonHocController@delete')->name('deleteMonHoc')->middleware('checkloginadmin::class');
+    });
 
     //Loai Diem
     Route::name('loaidiemManage.')->group(function () {
-        Route::get('/loaidiem', 'App\Http\Controllers\LoaiDiemController@index')->name('indexLoaiDiem');
-        Route::get('/loaidiem/create', 'App\Http\Controllers\LoaiDiemController@create')->name('createLoaiDiem');
-        Route::post('/loaidiem/store', 'App\Http\Controllers\LoaiDiemController@store')->name('storeLoaiDiem');
-        Route::get('/loaidiem/edit/{maloaidiem}', 'App\Http\Controllers\LoaiDiemController@edit')->name('editLoaiDiem');
-        Route::post('/loaidiem/update', 'App\Http\Controllers\LoaiDiemController@update')->name('updateLoaiDiem');
-        Route::delete('/loaidiem/delete/{maloaidiem}', 'App\Http\Controllers\LoaiDiemController@delete')->name('deleteLoaiDiem');
-    })->middleware('checkloginadmin::class');
+        Route::get('/loaidiem', 'App\Http\Controllers\LoaiDiemController@index')->name('indexLoaiDiem')->middleware('checkloginadmin::class');
+        Route::get('/loaidiem/create', 'App\Http\Controllers\LoaiDiemController@create')->name('createLoaiDiem')->middleware('checkloginadmin::class');
+        Route::post('/loaidiem/store', 'App\Http\Controllers\LoaiDiemController@store')->name('storeLoaiDiem')->middleware('checkloginadmin::class');
+        Route::get('/loaidiem/edit/{maloaidiem}', 'App\Http\Controllers\LoaiDiemController@edit')->name('editLoaiDiem')->middleware('checkloginadmin::class');
+        Route::post('/loaidiem/update', 'App\Http\Controllers\LoaiDiemController@update')->name('updateLoaiDiem')->middleware('checkloginadmin::class');
+        Route::delete('/loaidiem/delete/{maloaidiem}', 'App\Http\Controllers\LoaiDiemController@delete')->name('deleteLoaiDiem')->middleware('checkloginadmin::class');
+    });
 
     // Quan ly diem
     Route::name('diemManage.')->group(function () {
-        Route::get('/diem/edit/hk{hocki}/{mahocsinh}/{mamonhoc}', 'App\Http\Controllers\DiemController@edit')->name('edit');
-        Route::post('/diem/update', 'App\Http\Controllers\DiemController@update')->name('update');
-    })->middleware('checklogincanbo::class');
+        Route::get('/diem/edit/hk{hocki}/{mahocsinh}/{mamonhoc}', 'App\Http\Controllers\DiemController@edit')->name('edit')->middleware('checklogincanbo::class');
+        Route::post('/diem/update', 'App\Http\Controllers\DiemController@update')->name('update')->middleware('checklogincanbo::class');
+    });
 });
